+++ sdk/spring/azure-spring-cloud-trace-sleuth/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 4.0.0-beta.1 (Unreleased)
 
 ### Features Added
+- Support http pipeline policy implemented by Spring Cloud Sleuth API.  ([#24192](https://github.com/Azure/azure-sdk-for-java/pull/24192))
 
 ### Breaking Changes